CHECK DE TOUT LES WARNINGS
Correction des comparaisons int / size_t
→ éviter les conflits signé/non-signé et les warnings -Wsign-compare

Ajout de conversions explicites static_cast<int>
→ empêcher les pertes silencieuses lors des conversions size_t → int

Ajout de conversions explicites static_cast<float>
→ éviter les promotions implicites int → double dans les calculs d’angles

Correction de l’utilisation d’une variable hors scope (i)
→ supprimer une erreur de compilation due à une variable inexistante

Suppression d’une boucle redondante dans afficher_modal
→ améliorer la cohérence du code et éviter les duplications inutiles

Correction des affectations x = size_t dans le chargement de carte
→ garantir des conversions contrôlées et sûres sur les coordonnées

Uniformisation des types dans les calculs de visibilité
→ assurer un comportement défini et éliminer les warnings liés aux conversions

+++ maj.hpp
@@ -157,13 +157,13 @@ void mettre_a_jour_visibilite(Jeu& jeu) {
 
     for (int i = 0; i < nb_branches; i++) {
         // calcul du vecteur direction du rayon
-        float angle = (2.0f * M_PI * i) / nb_branches;
+        float angle = (2.0f * static_cast<float>(M_PI) * static_cast<float>(i)) / static_cast<float>(nb_branches);
         float dx = cosf(angle);
         float dy = sinf(angle);
 
         // position initiale du rayon centrée sur la case
-        float cx = px + 0.5f;
-        float cy = py + 0.5f;
+        float cx = static_cast<float>(px) + 0.5f;
+        float cy = static_cast<float>(py) + 0.5f;
 
         int prev_ix = (int)floor(cx);
         int prev_iy = (int)floor(cy);
